+++ services/catalog/service.go
@@ -3,7 +3,7 @@ package catalog
 import "context"
 
 type Service interface {
-	CreateProduct(ctx context.Context, name, description string, price float32) (Product, error)
+	CreateProduct(ctx context.Context, name, description string, price int64) (Product, error)
 	GetProductByID(ctx context.Context, id string) (Product, error)
 	GetProducts(ctx context.Context, offset, limit int32) ([]Product, error)
 	GetProductsByIDs(ctx context.Context, ids []string) ([]Product, error)
@@ -18,7 +18,7 @@ func NewService(r Repository) Service {
 	return &service{r}
 }
 
-func (s *service) CreateProduct(ctx context.Context, name, description string, price float32) (Product, error) {
+func (s *service) CreateProduct(ctx context.Context, name, description string, price int64) (Product, error) {
 	return s.repository.CreateProduct(ctx, productDocument{Name: name, Description: description, Price: price})
 }
